Don't set TokenWithdrawer field on Binance withdrawal receipt (#126)
The TokenWithdrawer field that was added to withdrawal receipts to support BEP2 withdrawals to Binance is redundant, and this has already led to bugs where this field was set incorrectly. The TokenWithdrawer on the receipt is meant to contain the DAppChain address of the withdrawer, however it's trivial to obtain this address because the receipt itself is associated with a specific DAppChain account. Therefore when the tg:v1.5 feature flag is enabled the TokenWithdrawer field will no longer be set.

This also adds a sanity check to ensure current withdrawer address is known in UpdateWithdrawalReceipt, though it's very unlikely that foreignAccount.CurrentWithdrawer will be nil at the point where it is unmarshalled by UpdateWithdrawalReceipt.